basic enemy chasing AI + removing of delta across game
Basic enemy now chases the player like a rabid dog. The main problem in the previous chase logic was the normalized in the first edited line of the script. To make the chasing more smooth, I lerp the steering force and MIN and MAX the chase speed. The delta was used across the game, but move and slide already uses delta internally so it was bad math.